generalized method for extracting lower dimension features of mesh
A funcition for generating the mesh edges was generated as that may be useful during the optimzation routine.
The meshFace function was also re-written to be the same style and will possibly be implemented as a get the M
dimension substructure of an N dimensional mesh later. we will see.