major updates to mesh conditioning
this commit adds a semi-recursive mesh conditioning algorithm for improving the quality of mesh cells with higly obtuse angles (>179.5deg). This prevents floating point arithmetic comparison errors in the
minimization and just improves the quality of the mesh all around. it has a number of dependencies that are associated with it. these routines have been fairly well tested but could use a little more maybe.
they are also far from optimized

we also add a routine to detect and delete duplicate faces as well as a routine for removing any cells with highly acute angles in which the adjacent mesh cells are merged to cover the gap.

acute->tighten the mesh
obtuse->relax the mesh