added vertex discrete optimization for mesh
Rather than optimize across the faces of each mesh using fmincon, we can also just test the vertices of the triangles in the mesh and consider those only.

If we do so, then we can optimize in one of two ways: we can use the same type of algorithm as we do in the fminmesh mode to traverse from mesh cell to mesh cell
or we can do a direct search using min to produce the global optimum among all of the mesh cells. It remains to be seen how these can be implemented and how fast each will be in the case
of a optimization of multiple feet at once.

Addresses Issue #1.