MAJOR - added functionality for external penalty function on mesh domain
This update allows an external penalty function to be used to enforce the mesh domain requirement.
This means that we can use bounded or unbounded optimization to solve the problem.

closestVertex calculates the closest vertex to a point and is useful in some analysis
distanceFromEdge is used to calculate the penalty weighting of the objective function
xyMeshIntercept maps xyPoints to the mesh domain as the domain is a nearly-flat surfaces (homeomorphic to plane)
meshPenalty takes an xy coorindate and computes the penalty function from the original objective function to force convergence into the mesh domain

This is a fundamental change in approach to Issue #1.