added combined constraints for leg crossing and stability threshold
A constraint function was added that allows the distance between the closest pair of legs to be specified as a
constraint as well as the minimum stability in a shuffle. Currently the minimum stability is the initial stability and the
leg distance is 0.1 units or about one tenth of the leg length.